Section 9: Authorization

+++ backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post("", checkauth, multer({storage: storage}).single("image") ,(req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
 
   console.log(post);
@@ -90,12 +91,18 @@ router.put("/:id", checkauth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Update successful!'})
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.modifiedCount > 0) {
+      res.status(200).json({message: 'Update successful!'});
+    }
+    else {
+      res.status(401).json({message: 'Not Authorised!'})
+    }
+
 
   })
 });
@@ -112,9 +119,15 @@ router.get("/:id", (req, res, next) => {
 
 
 router.delete("/:id", checkauth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post Deleted"});
+    if (result.deletedCount > 0) {
+      res.status(200).json({message: 'Deleted successful!'});
+    }
+    else {
+      res.status(401).json({message: 'Not Authorised!'})
+    }
+    // res.status(200).json({ message: "Post Deleted"});
   })
 });
